Refactor: 날짜 생성 함수 분리
