+++ src/utils/formatDate.js
@@ -1,0 +1,4 @@
+export function formatDate(value) {
+	const date = new Date(value);
+	return `${date.getFullYear()}. ${date.getMonth() + 1}. ${date.getDate()}`;
+}
